fix refactoring folder structure
