+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import logo from './logo.svg';
 import './App.css';
-import Login from './components/Login';
+import Login from './routes/Login';
 
 function App() {
   return (
